delete open_all_fd.c
